Added my Loop shape, also added tunnels detection.
After hit button detect tunnels it put point on place where wires are crossing boundary box of raft object. I calculate crossing of each segment of wire with boundary box raft object to get proper position. Also based on direction which wire is crossing boundary box is this point pushed inside raft object so when raft object is moving, these points are inside it, so they are moving also.
Raft object means, that if object is inside boundary of it it's moving together with it, but there is no internal signal connection btw these two objects.