Added tunnel generation, after hit "Detect Tunnels" editor add tunnels to loop at each connection intersection with loop.
I wanted to implement it as hybrid port, but for now don't know how to set wires direction from hybrid port, now they are always going in one direction doesn't matter if to port is wired input or output. It should be also changing when tunnel is on left, right, top or bottom side.
So now it's implemented as own figure, which has input and output and is placed close to loop border and connection which is going through loop border is splitted and reconnected to tunnel.