IMRPOVED DOCUMENTATION, start working on fucntion to get rid of z-order of while loops.
As written I figure out how to use @param tag, so I write documentation to most of the functions. And now even I start to oriented in my code. I see that one way from this interpreter javascript hell is to write documentation and use it as tool, becasue in this event driven languages where logic is similar to real world I cannot hold everything in my head and trust in documentation and my ability to read my code, because sometimes I can be little drunk when I write it and then I don't remember all implementation issues which I tried to solved it so I have to write them down somewhere.

This is just so long becuase today I was working on this after little longer break and still I have to solve some administrative issues here in Czech republic, but hopefully everything will be good. :)