Added implementation of Constants, for now basic but it's run, their values are copied into generated code without problem. There are still like alway some esoteric problems.
As written above, I implemented constants based on exmaple from ndreas Hertz with node with editable label. That's enough for me for now. Just put there node, it's background is based on color of it's datatype and double click on middle label and edit it's value. This would be transcripted directly as it is to the generated source code, so there is no protection for wrong input.