Wires corrected, there is button after load wires will be connected bad to tunnels, this is going through them and correct it because I added some info about it to userData of each wire.
During serialization is checked if wire is connected to the tunnel (by its source or target, depend how user drag wire, when click by mouse it created source of wire during drag and after drop it created target) and store this information into:
    userData.sourceTunnel
    userData.targetTunnel
    userData.sourcePortName
    userData.targetPortName

I have to done function to button becasue I have to probably wait until all will be readed to be able play with canvas objects and so. So now it's running somehow.