Added button to see selected loop aboard nodes, still problem with multilayered not moving when part of loop due its elements are added to appCanvas rather then this.
If subelements are added to canvas when elements are putted into layers they become layer parts. When added to this, then structure is moving rught with parent loop but elements are not becoming part of layers when putted over layer.
So need to think how to make movable, even detectable multilayer structure, because even loops are not seeing these multilayered nodes :(