CORRECTED MULTILAYERED figure, now it's running as it should be, it's saved correctly into file, there is no rectangle to protect it during moving and seems it's loading correcty because when StrongComposite is used every figure knows to which figure it belongs. This is added in Multilayered3.js
I switch Loop2.js to use StrongComposite, so now all loops and everything behave like jailhouse due to be make this running as it should be. Lot of things simplified.
I removed unecessary parts from Multilayered.js and saved it as Multilayered3.js.

While loop implemented as jailhouse is in file WhileLayer.js

NEED TO DO LOADING CONNECTIONS in Utils.js, they should be removed from initial list and created after all figures are created to have ports to can connect to them, because ports must be created in canvas to be able create connection programaticaly.